:recycle: refactor: post get 요청으로 전환
