+++ src/app/api/board/detail/route.ts
@@ -1,12 +1,12 @@
-import { postBoardDetail } from '@/app/service/postRequests'
+import { getBoardDetail } from '@/app/service/getRequests'
 import { getCookie } from '@/app/utils/cookies'
 import { NextResponse } from 'next/server'
 
-export async function POST(request: Request) {
+export async function GET(request: Request):Promise<NextResponse> {
   let accessToken = getCookie(request, 'accessToken')
   const { searchParams } = new URL(request.url)
   const id = searchParams.get('id') || ''
-  const data = await postBoardDetail(accessToken, id)
+  const data = await getBoardDetail(accessToken, id)
 
   return NextResponse.json(data)
 }
