+++ src/app/components/board/containers/BoardDetailContainer.tsx
@@ -48,12 +48,6 @@ const BoardDetailContainer = ({ nickname }: BoardDetailContainerProps) => {
       const fetchData = async () => {
         const response = await fetch(
           `${process.env.NEXT_PUBLIC_NEXT_SERVER}/api/board/detail?id=${selectedId}`,
-          {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-          },
         )
         const resData = await response.json()
         setData(resData.result.boardInfo)
